# tacocloud/tacocloud-api/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Taco Cloud API
  version: 1.0.0
  description: >-
    Public v1 contract. Legacy /api routes remain available until 2027-12-31 and return
    Deprecation, Sunset and Link headers pointing to the equivalent /api/v1 route.
servers:
  - url: /api/v1
...
paths:
  /ingredients:
    get:
      tags: [Ingredients]
      operationId: listIngredients
      responses:
        '200':
          $ref: '#/components/responses/IngredientList'
    post:
      tags: [Ingredients]
      operationId: createIngredient
      security: [{cookieAuth: []}]
      requestBody:
        $ref: '#/components/requestBodies/Ingredient'
      responses:
        '201':
          $ref: '#/components/responses/Ingredient'
        '400':
          $ref: '#/components/responses/Problem'
  /ingredients/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Ingredients]
      operationId: getIngredient
      responses:
        '200': {$ref: '#/components/responses/Ingredient'}
        '404': {$ref: '#/components/responses/Problem'}
    put:
      tags: [Ingredients]
      operationId: replaceIngredient
      security: [{cookieAuth: []}]
      requestBody: {$ref: '#/components/requestBodies/Ingredient'}
      responses:
        '200': {$ref: '#/components/responses/Ingredient'}
        '400': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Ingredients]
      operationId: deleteIngredient
      security: [{cookieAuth: []}]
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '404': {$ref: '#/components/responses/Problem'}
...
  /tacos/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Tacos]
      operationId: getTaco
      responses:
        '200': {$ref: '#/components/responses/Taco'}
        '404': {$ref: '#/components/responses/Problem'}
...
  /orders:
    post:
      tags: [Orders]
      operationId: createOrder
      security: [{cookieAuth: []}]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OrderCreateRequest'}
            example:
              deliveryName: Daniel
              deliveryStreet: Calle 1
              deliveryCity: Guadalajara
              deliveryState: Jalisco
              deliveryZip: '44100'
              paymentMethodId: pm_123
              items:
                - quantity: 2
                  taco:
                    name: Taco especial
                    ingredientIds: [FLTO, GRBF]
      responses:
        '201':
          description: Created order or the original response for an idempotent retry
          headers:
            X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/OrderResponse'}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
        '422': {$ref: '#/components/responses/Problem'}
  /orders/fromEmail:
    post:
      tags: [Orders]
      operationId: createOrderFromEmail
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/EmailOrderRequest'}
      responses:
        '201': {$ref: '#/components/responses/Order'}
        '400': {$ref: '#/components/responses/Problem'}
...
  /orders/{orderId}:
    parameters:
      - {name: orderId, in: path, required: true, schema: {type: string}}
    patch:
      tags: [Orders]
      operationId: patchOrder
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OrderPatchRequest'}
      responses:
        '200': {$ref: '#/components/responses/Order'}
        '400': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
    put:
      tags: [Orders]
      operationId: replaceOrder
      security: [{cookieAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/OrderReplaceRequest'}
      responses:
        '200': {$ref: '#/components/responses/Order'}
        '400': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Orders]
      operationId: deleteOrder
      security: [{cookieAuth: []}]
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '404': {$ref: '#/components/responses/Problem'}
  /orders/{id}/status:
    parameters:
      - $ref: '#/components/parameters/Id'
    patch:
      tags: [Orders]
      operationId: updateOrderStatus
      security: [{cookieAuth: []}]
      requestBody: {$ref: '#/components/requestBodies/OrderStatus'}
      responses:
        '200': {$ref: '#/components/responses/OrderDetail'}
        '400': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
...
  /users/me/favorites:
    get:
      tags: [Favorites]
      operationId: listFavorites
      security: [{cookieAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
      responses:
        '200': {$ref: '#/components/responses/FavoritePage'}
  /users/me/favorites/{tacoId}:
    parameters:
      - {name: tacoId, in: path, required: true, schema: {type: string}}
    put:
      tags: [Favorites]
      operationId: addFavorite
      security: [{cookieAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/Taco'}
        '404': {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Favorites]
      operationId: removeFavorite
      security: [{cookieAuth: []}]
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
...
  /users/me/orders/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Orders]
      operationId: getMyOrder
      security: [{cookieAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/OrderDetail'}
        '404': {$ref: '#/components/responses/Problem'}
  /kitchen/queue:
    get:
      tags: [Kitchen]
      operationId: kitchenQueue
      security: [{cookieAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/KitchenOrderList'}
  /kitchen/orders/claim:
    post:
      tags: [Kitchen]
      operationId: claimKitchenOrder
      security: [{cookieAuth: []}]
      responses:
        '200': {$ref: '#/components/responses/KitchenOrder'}
        '404': {$ref: '#/components/responses/Problem'}
  /kitchen/orders/{id}/status:
    parameters:
      - $ref: '#/components/parameters/Id'
    patch:
      tags: [Kitchen]
      operationId: updateKitchenOrderStatus
      security: [{cookieAuth: []}]
      requestBody: {$ref: '#/components/requestBodies/OrderStatus'}
      responses:
        '200': {$ref: '#/components/responses/KitchenOrder'}
        '409': {$ref: '#/components/responses/Problem'}
...
  /admin/orders:
    get:
      tags: [Administration]
      operationId: listAllOrders
      security: [{cookieAuth: []}]
      parameters:
        - {name: userId, in: query, schema: {type: string}}
        - {name: status, in: query, schema: {$ref: '#/components/schemas/OrderStatus'}}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/Size'}
      responses:
        '200': {$ref: '#/components/responses/OrderPage'}
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: {type: string}
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 0, default: 0}
    Size:
      name: size
      in: query
      schema: {type: integer, minimum: 1, maximum: 50, default: 20}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 8, maxLength: 100, pattern: '^[A-Za-z0-9._:-]+$'}
  headers:
    CorrelationId:
      description: Correlation identifier accepted from the request or generated by the server
      schema: {type: string}
  requestBodies:
    Ingredient:
      required: true
      content:
        application/json:
          schema: {$ref: '#/components/schemas/IngredientRequest'}
...
    OrderStatus:
      required: true
      content:
        application/json:
          schema: {$ref: '#/components/schemas/OrderStatusUpdateRequest'}
  responses:
    NoContent:
      description: Operation completed without a response body
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
    Problem:
      description: RFC 9457 compatible API error
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/ApiProblem'}
          example:
            type: urn:tacocloud:problem:validation-error
            title: Request validation failed
            status: 400
            detail: One or more fields are invalid
            instance: /api/v1/orders
            code: VALIDATION_ERROR
            violations: [{field: deliveryZip, reason: must contain exactly 5 digits}]
    Ingredient:
      description: Ingredient response
      headers:
        X-Correlation-Id: {$ref: '#/components/headers/CorrelationId'}
      content: {application/json: {schema: {$ref: '#/components/schemas/IngredientResponse'}}}
    IngredientList:
      description: Ingredient list
      content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/IngredientResponse'}}}}
    Taco:
      description: Taco response
      content: {application/json: {schema: {$ref: '#/components/schemas/TacoResponse'}}}
...
    Order:
      description: Order response
      content: {application/json: {schema: {$ref: '#/components/schemas/OrderResponse'}}}
    OrderDetail:
      description: Detailed order response
      content: {application/json: {schema: {$ref: '#/components/schemas/OrderDetailResponse'}}}
    OrderPage:
      description: Order history page
      content: {application/json: {schema: {$ref: '#/components/schemas/OrderPageResponse'}}}
...
    FavoritePage:
      description: Favorite taco page
      content: {application/json: {schema: {$ref: '#/components/schemas/FavoritePageResponse'}}}
    KitchenOrder:
      description: Kitchen order
      content: {application/json: {schema: {$ref: '#/components/schemas/KitchenOrderResponse'}}}
    KitchenOrderList:
      description: Kitchen queue
      content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/KitchenOrderResponse'}}}}
...
  schemas:
    IngredientRequest:
      type: object
      additionalProperties: false
      required: [name, type]
      properties:
        name: {type: string, minLength: 2, maxLength: 50}
        type: {$ref: '#/components/schemas/IngredientType'}
    IngredientResponse:
      type: object
      required: [id, name, type, unitPrice, available]
      properties:
        id: {type: string}
        name: {type: string}
        type: {$ref: '#/components/schemas/IngredientType'}
        unitPrice: {type: number}
        available: {type: boolean}
        dietaryTags: {type: array, items: {type: string}}
        allergens: {type: array, items: {type: string}}
        spiceLevel: {type: string}
    IngredientType:
      type: string
      enum: [WRAP, PROTEIN, VEGGIES, CHEESE, SAUCE]
...
    TacoResponse:
      type: object
      required: [id, name, ingredients]
      properties:
        id: {type: string}
        name: {type: string}
        createdAt: {type: string, format: date-time}
        ingredients: {type: array, items: {$ref: '#/components/schemas/IngredientResponse'}}
        classification: {$ref: '#/components/schemas/TacoClassificationResponse'}
...
    TacoClassificationResponse:
      type: object
      properties:
        dietaryTags: {type: array, items: {type: string}}
        allergens: {type: array, items: {type: string}}
        spiceLevel: {type: string}
...
    OrderTacoRequest:
      type: object
      additionalProperties: false
      required: [name, ingredientIds]
      properties:
        name: {type: string, minLength: 5, maxLength: 50}
        ingredientIds: {type: array, minItems: 1, maxItems: 10, items: {type: string}}
    OrderItemRequest:
      type: object
      additionalProperties: false
      required: [taco, quantity]
      properties:
        taco: {$ref: '#/components/schemas/OrderTacoRequest'}
        quantity: {type: integer, minimum: 1}
    OrderCreateRequest:
      type: object
      additionalProperties: false
      required: [deliveryName, deliveryStreet, deliveryCity, deliveryState, deliveryZip, paymentMethodId, items]
      properties:
        deliveryName: {type: string, maxLength: 100}
        deliveryStreet: {type: string, maxLength: 120}
        deliveryCity: {type: string, maxLength: 80}
        deliveryState: {type: string, maxLength: 80}
        deliveryZip: {type: string, pattern: '^\d{5}$'}
        paymentMethodId: {type: string}
        couponCode: {type: string, maxLength: 40}
        items: {type: array, minItems: 1, maxItems: 20, items: {$ref: '#/components/schemas/OrderItemRequest'}}
    OrderResponse:
      type: object
      required: [id, status, userId, deliveryName, deliveryStreet, deliveryCity, deliveryState, deliveryZip, subtotal, discount, total, currency, items]
      properties:
        id: {type: string}
        placedAt: {type: string, format: date-time}
        status: {$ref: '#/components/schemas/OrderStatus'}
        userId: {type: string}
        deliveryName: {type: string}
        deliveryStreet: {type: string}
        deliveryCity: {type: string}
        deliveryState: {type: string}
        deliveryZip: {type: string}
        paymentMethod: {$ref: '#/components/schemas/PaymentMethodResponse'}
        subtotal: {type: number}
        discount: {type: number}
        total: {type: number}
        currency: {type: string, example: MXN}
        couponCode: {type: string, nullable: true}
        items: {type: array, items: {$ref: '#/components/schemas/OrderItemResponse'}}
    OrderItemResponse:
      type: object
      properties:
        taco: {$ref: '#/components/schemas/TacoResponse'}
        quantity: {type: integer}
        unitPrice: {type: number}
        lineTotal: {type: number}
    OrderPatchRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        deliveryName: {type: string, maxLength: 100}
        deliveryStreet: {type: string, maxLength: 120}
        deliveryCity: {type: string, maxLength: 80}
        deliveryState: {type: string, maxLength: 80}
        deliveryZip: {type: string, pattern: '^\d{5}$'}
    OrderReplaceRequest:
      type: object
      additionalProperties: false
      required: [deliveryName, deliveryStreet, deliveryCity, deliveryState, deliveryZip, tacos]
      properties:
        deliveryName: {type: string}
        deliveryStreet: {type: string}
        deliveryCity: {type: string}
        deliveryState: {type: string}
        deliveryZip: {type: string, pattern: '^\d{5}$'}
        tacos: {type: array, minItems: 1, maxItems: 20, items: {$ref: '#/components/schemas/OrderReplacementTacoRequest'}}
    OrderReplacementTacoRequest:
      type: object
      required: [name, ingredients]
      properties:
        name: {type: string, minLength: 5}
        ingredients: {type: array, minItems: 1, items: {$ref: '#/components/schemas/IngredientResponse'}}
    OrderStatus:
      type: string
      enum: [CREATED, PLACED, ACCEPTED, PREPARING, READY, OUT_FOR_DELIVERY, DELIVERED, CANCELLED]
    OrderStatusUpdateRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status: {$ref: '#/components/schemas/OrderStatus'}
        reason: {type: string, maxLength: 200}
...
    OrderDetailResponse:
      allOf:
        - $ref: '#/components/schemas/OrderResponse'
        - type: object
          properties:
            statusHistory: {type: array, items: {type: object}}
    OrderPageResponse:
      type: object
      properties:
        content: {type: array, items: {$ref: '#/components/schemas/OrderResponse'}}
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer}
...
    EmailOrderRequest:
      type: object
      additionalProperties: false
      required: [email, tacos]
      properties:
        email: {type: string, format: email}
        tacos:
          type: array
          minItems: 1
          maxItems: 20
          items:
            type: object
            required: [name, ingredients]
            properties:
              name: {type: string, minLength: 5, maxLength: 50}
              ingredients: {type: array, minItems: 1, maxItems: 10, items: {type: string}}
...
    FavoritePageResponse:
      type: object
      properties:
        content: {type: array, items: {$ref: '#/components/schemas/TacoResponse'}}
        page: {type: integer}
        size: {type: integer}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer}
    KitchenOrderResponse:
      type: object
      properties:
        id: {type: string}
        placedAt: {type: string, format: date-time}
        status: {$ref: '#/components/schemas/OrderStatus'}
        stationId: {type: string}
        cookId: {type: string, nullable: true}
        acceptedAt: {type: string, format: date-time, nullable: true}
        estimatedPrepMinutes: {type: integer}
        items: {type: array, items: {type: object}}
...
    PaymentMethodResponse:
      type: object
      properties:
        paymentMethodId: {type: string}
        brand: {type: string}
        last4: {type: string, pattern: '^\d{4}$'}
...
    ApiProblem:
      type: object
      required: [type, title, status, detail, instance, code, violations]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string}
        instance: {type: string, format: uri-reference}
        code: {type: string}
        violations: {type: array, items: {$ref: '#/components/schemas/ApiViolation'}}
    ApiViolation:
      type: object
      required: [field, reason]
      properties:
        field: {type: string}
        reason: {type: string}